refactor(CashAmountParserTeat): ParseTest로 리펙토링
CashAmountParser가 Parser로 변경됨에 따라 그에 맞는 단위 테스트로 변경한다.
